v0.3.1024 - /mep leaves client.ts
SCALE-SEAM ⑲. Seven MEP methods (summary, connectivity, sizing, sprinkler,
fittings, class rollup, model-extract) move to api/mep.ts. Four regions.
connectMep / addMepFitting stay on editIfc. client.ts 3,304 → 3,243.
