+++ apps/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aec/web",
-    "version": "0.3.1023",
+    "version": "0.3.1024",
   "private": true,
   "type": "module",
   "engines": {
